Changed all config sys V ipc stuff to use the new mm module.
Everything compiles cleanly, but it's not yet tested. All password
stuff is unchanged (the cli_config structure is the same). Now cfg
is a pointer to struct mm_private, and the cli_config structure is
mapped to the static part of the mm area. CFG is a macro that
converts the (globally available) cfg pointer into a struct
cli_config pointer. Thus all old cfg-> references were changed to
CFG-> references.

Signed-off-by: Radu Rendec <radu.rendec@ines.ro>