Fix glitches with xen integration script (vif-lisa).
Added "silent mode" support to swcfgload with tagged interface:
When a vm is created for the first time, there is no tag configuration
script in /etc/lisa/tags, so just setup the interface (add to switch
and tag it) and silently ignore the configuration file.

Added "silent mode" support to swctl del:
On silent mode, ignore the ioctl() exit status and delete
the tag anyway. This is useful for Xen integration: Xen
first unregisters vifX.0, and then calls the vif-lisa
script. The netdevice is removed from the switch immediately
(via netdevice notification handler in switch kernel module,
but the interface tag must be deleted later, via userspace
brctl).

Signed-off-by: Radu Rendec <radu.rendec@ines.ro>