Make inline help work
Inline help (the '?' key) is fully functional and (hopefully) covers
all possible cases. We still need completion (the 'tab' key) and
command execution (the 'enter' key).

Also made a few changes to the tokenize() design and adjusted the
README file accordingly. Changes are listed below.

Change 1: The return value condition was "number of matches > 1 &&
whitespace after the token". Since calling functions can determine
themselves if "matches > 1", I only kept the second condtition and
thus changed the meaning of the return value into "whitespace after
the token". This approach simplifies the logic in the calling
functions.

Change 2: If no matches are found, determine the maximum initial
substring of the token that would generate at least 1 match. This
helps identifying the exact point of the parse error (useful to
display the exact position of '^' in the "invalid input detected..."
message.

Change 3: allow "matching" against an empty subtree; this eases
detection of junk after a token that matches the deepest subtree
node.

Signed-off-by: Radu Rendec <radu.rendec@ines.ro>