Change menu node subtree to array of pointers to other nodes
Menu node subtree was an array of structures (other nodes). This way
it was impossible to declare an isolated node as variable, and then
use it as subnode in various other nodes.

Also cleaned up the "show mac-address-table" subnodes by declaring
them as isolated nodes and properly linking them. The same procedure
will be used for output modifiers (pipe), but code must be adapted
and moved to a propper place.

Signed-off-by: Radu Rendec <radu.rendec@ines.ro>