Fix completion with custom tokenizers
Completion was automatically suppressed for nodes with custom
tokenizers. This was done to prevent completion on generic nodes
such as "WORD", "LINE", interface and vlan numbers etc.

Still, completion must be active even with custom tokenizers. To
achieve this, add the suppress_completion flag to struct
rlshell_context. The flag is automatically reset by the completion
handler and set by custom tokenizers that need to prevent the
completion from taking place.

Signed-off-by: Radu Rendec <radu.rendec@ines.ro>