Fully implement interface addition / selection
All possible cases for "interface <type> <arg>" are (hopefully) covered.
The code should automatically adapt and still take the appropriate
decisions if interfaces are renamed. It's should also be easy to extend
if other interface types are implemented.

Command root is still *not* changed to "config-if" even on successful
interface addition / selection.

Signed-off-by: Radu Rendec <radu.rendec@ines.ro>