Partially accomodate kernel module changes in cli1
Renamed struct net_switch_ioctl_arg to swcfgreq and .if_name to
.ifindex. This makes the code build (with warnings, because .if_name
was "char *" and .ifindex is "int").

Still, the code is *BROKEN* since the userspace part treats the
field as char pointer to interface *name*, while the kernel part
expects it to be an interface *index*. Probably this won't ever be
fixed since all work is now done on cli2, which, eventually, will
completely replace cli1.

Signed-off-by: Radu Rendec <radu.rendec@ines.ro>