Partially accomodate kernel module changes in cli, cdpd, lib and swctl
For cli, changed ioctl subcommands to accomodate kernel module unified
port and vif disable/enable commands.

For cdpd, lib and swctl, renamed struct net_switch_ioctl_arg to
swcfgreq and .if_name to .ifindex.

Still, the code is *BROKEN* since the userspace part treats the
field as char pointer to interface *name*, while the kernel part
expects it to be an interface *index*.

Signed-off-by: Radu Rendec <radu.rendec@ines.ro>