Config-if: implemented shutd / no shutd; sw mode access / sw mode trunk
Added cmd_ioctl_simple(), which is a handler for simple commands that
only need to do some SIOCSWCFG ioctl()'s. The trick is to use the .priv
member of struct menu_node to define an array of (pointers to) struct
swcfgreq. The handler simply looks at the last matched node's .priv
member and fires the ioctl() calls. Current interface index and vlan
are taken from cli context private data.

Signed-off-by: Radu Rendec <radu.rendec@ines.ro>