Make tokenizers aware of rlshell's completion
Some special nodes must not be completed by rlshell, for instance "WORD"
or "LINE". This was accomplished by having the tokenizer let rlshell
know that completion must not be done on that node.

However, there are special cases where the tokenizer results depend on
whether the tokenizer is used for completion or not. Example:

(config-if)# switchport trunk allowed vlan WORD
                                           add
                                           all
                                           except
                                           none
                                           remove

With a buffer like "switchport trunk allowed vlan r", the node above
must produce two matches ("WORD" and "remove") for inline help and
command execution. However, completion must never see "WORD" as a
match (since it cannot complete on it) and complete on "remove"
instead.

To achieve this, tokenizers must know if they are used for completion
or not. The "suppress_completion" member in "struct rlshell_context"
was replaced by "completion" which is now set by the caller (rlshell
completion) and examined by the tokenizer (and not vice-versa).

Signed-off-by: Radu Rendec <radu.rendec@ines.ro>