Change SW_SOCK_OPEN semantics
It is desirable that SW_SOCK_OPEN behaves like a function with return value.
This way the user will decide if aborting is the right thing to do or not.

Changed command handlers to check for the result of SW_SOCK_OPEN and return
the error code through the available cli mechanisms.

Signed-off-by: Ionut Nicu <ionut.nicu@ines.ro>