Change CDP_SESSION_OPEN semantics
It is desirable that the CDP_SESSION_OPEN macro behaves like a function with return value.
This way the user will decide if aborting is the right thing to do or not.

Changed command handlers to check for the result of CDP_SESSION_OPEN and return
the error code through the available cli mechanisms or warn the user (for example
when adding/removing interface from switch).

Signed-off-by: Ionut Nicu <ionut.nicu@ines.ro>