Short canonical interface names in "show vlan".
Moved canonical_if_name to if_generic.c and implemented short_if_name with the
same functionality for now. The main reason for separating behavior for long
canonical names and for short canonical names is that there may be different
cases when we add more interface types.

Also, ports are listed alphabetically now.