Fix userspace compilation on Ubuntu
Apparently on Ubuntu there's a difference when
the $(LD_FLAGS) are placed before or after the
'-o' switch. We observed linker errors (undefined
symbols) when $(LD_FLAGS) is placed before the
'-o' switch.

Also the libswitch.so library should be dinamically
loaded using the '-lswitch' linker flag instead
of compiling it like a regular object file.

Signed-off-by: Ionut Nicu <ionut.nicu@cloudbit.ro>