[CLI to sw_api] Replace SWCFG_GETIFTYPE ioctls.
issue #39
Replace mentioned ioctl with sw_ops->if_get_type. Note the (sw_ops->if_get_type)
notation as there is a macro with the same name. The macro still receives the
socket file descriptor as parameter in order to close the socket if an error
occurs, but it's not used anymore for an ioctl call.

Signed-off-by: Andreea Hodea <hodea_andreea@yahoo.com>