Redesign SW_API
Modified get_vlan_interfaces so that it will take an int** ifindexes instead of int* interfaces .
From now on backends must determine the number of interfaces in vlan and alloc it's own interface buffer and return it via the "ifindexes" field.
Made modifications in Bridge+8021q implementation.
Modification to Lisa backend should also be made.

Signed-off-by: Victor Duta <duta.victor.marin@gmail.com>